Simplify regex parser and ast
