+++ internal/regex/ast/ast_test.go
@@ -1,4 +1,4 @@
-package basic
+package ast
 
 import (
 	"testing"
